# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fashion Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sizes
          in: query
          schema:
            type: string
        - name: colors
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ── Orders ────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin sees all, customer sees own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── Cart ──────────────────────────────────────────────────────────────
...
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Wishlist ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ───────────────────────────────────────────────────────────
...
    ProductVariant:
      type: object
      required: [id, color, size, stockQuantity]
      properties:
        id:
          type: integer
        color:
          type: string
        size:
          type: string
        stockQuantity:
          type: integer
    ProductImage:
      type: object
      required: [id, imageUrl, isPrimary]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        isPrimary:
          type: boolean
    Product:
      type: object
      required: [id, nameEn, nameAr, price, categoryId, vendorId, active, createdAt]
      properties:
        id:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        descriptionEn:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        sku:
          type: ["string", "null"]
        featured:
          type: boolean
        active:
          type: boolean
        averageRating:
          type: number
        reviewCount:
          type: integer
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [nameEn, nameAr, price, categoryId]
      properties:
        nameEn:
          type: string
        nameAr:
          type: string
        descriptionEn:
          type: string
        descriptionAr:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        sku:
          type: string
        featured:
          type: boolean
        active:
          type: boolean
        images:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/VariantInput"
    ProductUpdate:
      type: object
      properties:
        nameEn:
          type: string
        nameAr:
          type: string
        descriptionEn:
          type: string
        descriptionAr:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        sku:
          type: string
        featured:
          type: boolean
        active:
          type: boolean
    VariantInput:
      type: object
      required: [color, size, stockQuantity]
      properties:
        color:
          type: string
        size:
          type: string
        stockQuantity:
          type: integer
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Order ──────────────────────────────────────────────────────────
    OrderItem:
      type: object
      required: [id, productId, productNameEn, productNameAr, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productNameEn:
          type: string
        productNameAr:
          type: string
        imageUrl:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
    Order:
      type: object
      required: [id, userId, totalPrice, paymentMethod, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        totalPrice:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        couponCode:
          type: ["string", "null"]
        discount:
          type: ["number", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [paymentMethod, items]
      properties:
        paymentMethod:
          type: string
        couponCode:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [productVariantId, quantity, price]
      properties:
        productVariantId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    OrderListResponse:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Cart ───────────────────────────────────────────────────────────
...
    Review:
      type: object
      required: [id, productId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
        comment:
          type: string

    # ── Wishlist ───────────────────────────────────────────────────────
...