# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fashion Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  userId: number;
  userName?: string;
  totalPrice: number;
  paymentMethod: string;
  status: string;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  discount?: number | null;
  items?: OrderItem[];
  createdAt: string;
}