# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fashion Store API
 * OpenAPI spec version: 0.1.0
 */
import type { VariantInput } from './variantInput';

export interface ProductInput {
  nameEn: string;
  nameAr: string;
  descriptionEn?: string;
  descriptionAr?: string;
  categoryId: number;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  sku?: string;
  featured?: boolean;
  active?: boolean;
  images?: string[];
  variants?: VariantInput[];
}